requirements.txt
asgiref==3.5.0
astroid==2.9.3
backports.zoneinfo==0.2.1
colorama==0.4.4
Django==4.0.2
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
phonenumbers==8.12.43
Pillow==9.0.1
platformdirs==2.5.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
tzdata==2021.5
wrapt==1.13.3
